Takes an alternate approach to exposing 'Text' fields, akin to how we do it with lazy search in NFT-SS1.
In this approach, you  just define properties with a 'Text' suffix  to the normal netsuite field. NSDAL looks for fields that end in 'Text' and treats them as GET/SET text.

See example.ts - we're using the native 'subsidiary' field on the customer record. That exists on the base class. So the Customer class defined in example.ts also wants the TEXT value for subsidiary so it just defines another property, decorated with the same type as the normal 'VALUE' subsidiary field. Only difference is the field name ends with 'Text'